Implement basic game renderer
