+++ app/src/main/java/com/example/scorekeeper/MainActivity.kt
@@ -49,6 +49,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.scorekeeper.game.GameStorage
+import com.example.scorekeeper.game.renderers.GameRenderer
 import com.example.scorekeeper.game.types.Game
 import com.example.scorekeeper.ui.theme.CAHAppTheme
 import com.example.scorekeeper.ui.theme.Purple500
@@ -154,7 +155,7 @@ fun AppMain(appViewModel: AppViewModel = viewModel()) {
         val ref = remember { Ref<Game>() }
 
         ref.value = appUiState.activeGame ?: ref.value
-        ref.value?.GamePage(appViewModel)
+        /*ref.value?.GamePage(appViewModel)*/
     }
 }
 
@@ -171,7 +172,8 @@ fun TitleText(title: String) {
 
 @Composable
 fun ColumnScope.GameList(appViewModel: AppViewModel) {
-    val games = GameStorage.getInstance(LocalContext.current).getGames()
+    val context = LocalContext.current
+    val games = GameStorage.getInstance(context).getGames()
         .collectAsState(initial = emptyList()).value
 
     LazyColumn(
@@ -184,7 +186,7 @@ fun ColumnScope.GameList(appViewModel: AppViewModel) {
         contentPadding = PaddingValues(12.dp)
 
     ) {
-        items(games) { game -> game.GetAsCard(appViewModel) }
+        items(games) { game -> GameRenderer.Card(game = game, appViewModel) }
     }
 }
