+++ app/src/main/java/com/example/scorekeeper/game/Player.kt
@@ -3,8 +3,12 @@ package com.example.scorekeeper.game
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Player(var name: String, var score: Int)
-
-fun List<String>.toPlayerList(): List<Player> {
-    return this.map { Player(it, 0) }.toList()
+data class Player(var name: String, var score: Int) {
+    companion object Factory {
+        fun toPlayerList(playerNames: List<String>): List<Player> {
+            return playerNames.map { Player(it, 0) }.toList()
+        }
+    }
 }
+
+
