--- app/src/main/java/com/example/scorekeeper/game/layout/GameView.kt
@@ -1,7 +1,0 @@
-package com.example.scorekeeper.game.layout
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun GameView(name: String) {
-}
