+++ app/src/main/java/com/example/scorekeeper/game/types/PointGame.kt
@@ -9,4 +9,15 @@ data class PointGame(
     override val players: List<Player>,
     override var playerSortOrder: SortingOrder,
     override var isComplete: Boolean
-) : Game()
+) : Game() {
+    companion object Factory {
+        fun new(name: String, playerNames: List<String>): PointGame {
+            return PointGame(
+                name,
+                Player.toPlayerList(playerNames),
+                SortingOrder.ALPHABETICAL,
+                false
+            )
+        }
+    }
+}
